pyproject.toml
[project]
name = "transpdfincn"
version = "0.1.0"
description = "翻译工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.102.0",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
]
